Memmodels and fences
All the memmodel-related code has been removed in this commit. Instead,
we just use the most conservative approach that is correct in any
machine: use fences always and add the .aq suffix to amo instructions.

This commit also uses the older `memory_barrier` optab instead of the
newer `mem_thread_fence`, which is too new and we don't have available.