Start a better guix.scm definiton for proper compilers
All the errors we had were me executing the compiler wrong. It is true
it has trouble finding things but if we call the compiler with -static
(we need glibc:static for that), we add the proper LIBRARY_PATH and the
-I flags, and we have binutils in the PATH it works properly.

In RISC-V things are complicated still... we need to port libgcc as
well as libatomic, before we use this new packaging for it.