riscv: uncomment va_start fixes segfault in genmddeps
genmddeps was segfaulting because it was missing a va_start function
definition: it accessed a non available part of the memory in the latest
argument.

This was working in the cross compiling setup as the compiler wasn't
building everything by itself but by the host compiler, which was way
more modern.