EMPTY: ucontext_t issues explanation:
Since 883312dc79806f513275b72502231c751c14ff72 gcc doesn't use `struct
ucontext` but we may need for the bootstrapping. If we are able to build
the other parts of the compiler that use that, we'll probably need to
revert this in the RISC-V part too.

I was able to build the RISC-V backend because we have the modern
approach rather than the old one, this might produce other problems
during the bootstrapping.

See also: a33c50fd788